Documented HomeFragment Class
Comments were added to allow for maintenance and step-by-step
understanding of the the processes and functionality of the
code; to make maintenance of the code for later android
revisions simpler.

This class is used by Home Activity and is important to keep
updated to ensure smooth operation of the application. Creates
the majority of the interface, with user profiles.