Documented HomeFragmentGridViewAdapter Class
Comments were added to allow for maintenance and step-by-step
understanding of the the processes and functionality of the
code; to make maintenance of the code for later android
revisions simpler.

This class is used by Home Fragment and is important to keep
updated since it acts as the rendering and controls system,
behind the Home Fragment.